Refactor unit test directories for better readability
