+++ api/src/tests/test.constants.ts
@@ -1,0 +1,3 @@
+import app from "../app";
+
+export const newApp = app;
